# openapi.yaml
openapi: 3.0.1
info:
  title: Sedna.Service.Requirements
  version: v1
paths:
  /api/Contexts:
    get:
      tags:
        - Contexts
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Context'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Context'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Context'
    post:
      tags:
        - Contexts
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/Context'
          application/json:
            schema:
              $ref: '#/components/schemas/Context'
          text/json:
            schema:
              $ref: '#/components/schemas/Context'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Context'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Context'
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
            text/json:
              schema:
                $ref: '#/components/schemas/Context'
  '/api/Contexts/{context_id}':
    get:
      tags:
        - Contexts
      parameters:
        - name: context_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Context'
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
            text/json:
              schema:
                $ref: '#/components/schemas/Context'
    put:
      tags:
        - Contexts
      parameters:
        - name: context_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/Context'
          application/json:
            schema:
              $ref: '#/components/schemas/Context'
          text/json:
            schema:
              $ref: '#/components/schemas/Context'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Context'
      responses:
        '200':
          description: Success
    delete:
      tags:
        - Contexts
      parameters:
        - name: context_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Context'
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
            text/json:
              schema:
                $ref: '#/components/schemas/Context'
...
  /v1/Requirements:
    get:
      tags:
        - Requirements
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Requirement'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Requirement'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Requirement'
    post:
      tags:
        - Requirements
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/Requirement'
          application/json:
            schema:
              $ref: '#/components/schemas/Requirement'
          text/json:
            schema:
              $ref: '#/components/schemas/Requirement'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Requirement'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Requirement'
            application/json:
              schema:
                $ref: '#/components/schemas/Requirement'
            text/json:
              schema:
                $ref: '#/components/schemas/Requirement'
  '/v1/Requirements/{requirement_id}':
    get:
      tags:
        - Requirements
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Requirement'
            application/json:
              schema:
                $ref: '#/components/schemas/Requirement'
            text/json:
              schema:
                $ref: '#/components/schemas/Requirement'
    put:
      tags:
        - Requirements
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/Requirement'
          application/json:
            schema:
              $ref: '#/components/schemas/Requirement'
          text/json:
            schema:
              $ref: '#/components/schemas/Requirement'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Requirement'
      responses:
        '200':
          description: Success
    delete:
      tags:
        - Requirements
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Requirement'
            application/json:
              schema:
                $ref: '#/components/schemas/Requirement'
            text/json:
              schema:
                $ref: '#/components/schemas/Requirement'
...
components:
  schemas:
    SourceType:
      type: object
      properties:
        sourceTypeId:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        source:
          type: array
          items:
            $ref: '#/components/schemas/Source'
          nullable: true
      additionalProperties: false
    Source:
      type: object
      properties:
        sourceId:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        sourceTypeId:
          type: integer
          format: int32
          nullable: true
        externalIdentifier:
          type: string
          nullable: true
        sourceType:
          allOf:
            - $ref: '#/components/schemas/SourceType'
          nullable: true
        requirement:
          type: array
          items:
            $ref: '#/components/schemas/Requirement'
          nullable: true
      additionalProperties: false
    TopicContext:
      type: object
      properties:
        topicContextId:
          type: integer
          format: int32
        topicId:
          type: integer
          format: int32
        contextId:
          type: integer
          format: int32
        context:
          allOf:
            - $ref: '#/components/schemas/Context'
          nullable: true
        topic:
          allOf:
            - $ref: '#/components/schemas/Topic'
          nullable: true
      additionalProperties: false
    TopicSearch:
      type: object
      properties:
        topicSearchId:
          type: integer
          format: int32
        topicId:
          type: integer
          format: int32
          nullable: true
        searchString:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isWholeWord:
          type: boolean
        topic:
          allOf:
            - $ref: '#/components/schemas/Topic'
          nullable: true
      additionalProperties: false
    Topic:
      type: object
      properties:
        topicId:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        preferredPhrasing:
          type: string
          nullable: true
        requirement:
          type: array
          items:
            $ref: '#/components/schemas/Requirement'
          nullable: true
        subjectTypeTopic:
          type: array
          items:
            $ref: '#/components/schemas/SubjectTypeTopic'
          nullable: true
        topicContext:
          type: array
          items:
            $ref: '#/components/schemas/TopicContext'
          nullable: true
        topicSearch:
          type: array
          items:
            $ref: '#/components/schemas/TopicSearch'
          nullable: true
      additionalProperties: false
    SubjectTypeTopic:
      type: object
      properties:
        subjectTypeTopicId:
          type: integer
          format: int32
        subjectTypeId:
          type: integer
          format: int32
        topicId:
          type: integer
          format: int32
        subjectType:
          allOf:
            - $ref: '#/components/schemas/SubjectType'
          nullable: true
        topic:
          allOf:
            - $ref: '#/components/schemas/Topic'
          nullable: true
      additionalProperties: false
    SubjectType:
      type: object
      properties:
        subjectTypeId:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        subject:
          type: array
          items:
            $ref: '#/components/schemas/Subject'
          nullable: true
        subjectTypeTopic:
          type: array
          items:
            $ref: '#/components/schemas/SubjectTypeTopic'
          nullable: true
      additionalProperties: false
    Subject:
      type: object
      properties:
        subjectId:
          type: integer
          format: int32
        externalIdentifier:
          type: string
          nullable: true
        subjectTypeId:
          type: integer
          format: int32
        tag:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        subjectType:
          allOf:
            - $ref: '#/components/schemas/SubjectType'
          nullable: true
        requirement:
          type: array
          items:
            $ref: '#/components/schemas/Requirement'
          nullable: true
      additionalProperties: false
    Requirement:
      type: object
      properties:
        requirementId:
          type: integer
          format: int32
        sourceId:
          type: integer
          format: int32
          nullable: true
        sourceText:
          type: string
          nullable: true
        sourceTextLocation:
          type: string
          nullable: true
        preferredPhrasing:
          type: string
          nullable: true
        topicId:
          type: integer
          format: int32
          nullable: true
        subjectId:
          type: integer
          format: int32
        isActive:
          type: boolean
          nullable: true
        reviewDt:
          type: string
          format: date-time
          nullable: true
        createdTs:
          type: string
          format: date-time
          nullable: true
        updatedTs:
          type: string
          format: date-time
          nullable: true
        createdUserId:
          type: integer
          format: int32
          nullable: true
        updatedUserId:
          type: integer
          format: int32
          nullable: true
        source:
          allOf:
            - $ref: '#/components/schemas/Source'
          nullable: true
        subject:
          allOf:
            - $ref: '#/components/schemas/Subject'
          nullable: true
        topic:
          allOf:
            - $ref: '#/components/schemas/Topic'
          nullable: true
        requirementContext:
          type: array
          items:
            $ref: '#/components/schemas/RequirementContext'
          nullable: true
      additionalProperties: false
    RequirementContext:
      type: object
      properties:
        requirementContextId:
          type: integer
          format: int32
        requirementId:
          type: integer
          format: int32
          nullable: true
        contextId:
          type: integer
          format: int32
          nullable: true
        context:
          allOf:
            - $ref: '#/components/schemas/Context'
          nullable: true
        requirement:
          allOf:
            - $ref: '#/components/schemas/Requirement'
          nullable: true
      additionalProperties: false
    Context:
      type: object
      properties:
        contextId:
          type: integer
          format: int32
        tag:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        requirementContext:
          type: array
          items:
            $ref: '#/components/schemas/RequirementContext'
          nullable: true
        topicContext:
          type: array
          items:
            $ref: '#/components/schemas/TopicContext'
          nullable: true
      additionalProperties: false